Switch to use Francois repository
