+++ mlreco/models/gnn/factories.py
@@ -16,7 +16,7 @@ def edge_model_dict():
     from . import edge_econv
     from . import edge_meta
     from . import dir_meta
-    from . import full_gnn
+    from . import modular_nnconv
 
     models = {
         "basic_attention" : edge_attention.BasicAttentionModel,
@@ -28,7 +28,7 @@ def edge_model_dict():
         "econv" : edge_econv.EdgeConvModel,
         "emeta" : edge_meta.EdgeMetaModel,
         "dir_meta" : dir_meta.EdgeMetaModel,
-        "full_gnn": full_gnn.FullGNN,
+        "modular_nnconv": modular_nnconv.NNConvModel,
     }
 
     return models
